test: consolidate activity ordering tests into pr.test.ts
Move the two ordering tests into the existing ActivityPRCommand describe
block, reusing createMockPullrequestsApi's activityPages option to drop
~120 lines of duplicated mock helpers.
